+++ src/xrdtiffoperations/imagemodel/ifd/fields/FieldTags.java
@@ -4,17 +4,17 @@ public class FieldTags {
 
     /////////// Constants ///////////////////////////////////////////////////////////////////
 
-    public static int IMAGE_WIDTH = 256;
-    public static int IMAGE_HEIGHT = 257;
-    public static int BITS_PER_SAMPLE = 258;
-    public static int COMPRESSION = 259;
-    public static int PHOTOMETRIC_INTERPRETATION = 262;
-    public static int STRIP_OFFSETS = 273;
-    public static int ORIENTATION = 274;
-    public static int ROWS_PER_STRIP = 278;
-    public static int STRIP_BYTE_COUNTS = 279;
-    public static int X_RESOLUTION = 282;
-    public static int Y_RESOLUTION = 283;
-    public static int RESOLUTION_UNIT = 296;
+    public static short IMAGE_WIDTH = 256;
+    public static short IMAGE_HEIGHT = 257;
+    public static short BITS_PER_SAMPLE = 258;
+    public static short COMPRESSION = 259;
+    public static short PHOTOMETRIC_INTERPRETATION = 262;
+    public static short STRIP_OFFSETS = 273;
+    public static short ORIENTATION = 274;
+    public static short ROWS_PER_STRIP = 278;
+    public static short STRIP_BYTE_COUNTS = 279;
+    public static short X_RESOLUTION = 282;
+    public static short Y_RESOLUTION = 283;
+    public static short RESOLUTION_UNIT = 296;
 
 }
